Refactor spørsmålsstilling angående arbeid
