+++ src/app/components/arbeidsforhold-detaljer/RedusertArbeidsforholdDetaljerPart.tsx
