+++ src/app/components/arbeidsforhold-detaljer/RedusertArbeidsforholdSNFDetaljerPart.tsx
