+++ src/app/components/arbeidsforhold-detaljer/info/arbeidsforholdInfo.tsx
@@ -1,23 +1,22 @@
 import Box from '@navikt/sif-common-core/lib/components/box/Box';
 import ExpandableInfo from '@navikt/sif-common-core/lib/components/expandable-content/ExpandableInfo';
 import intlHelper from '@navikt/sif-common-core/lib/utils/intlUtils';
+import { Element } from 'nav-frontend-typografi';
 import React from 'react';
 import { FormattedMessage, useIntl } from 'react-intl';
-import { Arbeidsform } from '../../types/PleiepengesøknadFormData';
-import { Element } from 'nav-frontend-typografi';
+import { Arbeidsform } from '../../../types/PleiepengesøknadFormData';
 
 interface Props {
     arbeidsform: Arbeidsform;
 }
 
-const ArbeidsformInfoSNFrilanser: React.FunctionComponent<Props> = ({ arbeidsform }) => {
+const ArbeidsformInfo: React.FunctionComponent<Props> = ({ arbeidsform }) => {
     const intl = useIntl();
     switch (arbeidsform) {
-        case Arbeidsform.varierende:
+        case Arbeidsform.fast:
             return (
-                <ExpandableInfo
-                    title={intlHelper(intl, 'snFrilanser.arbeidsforhold.arbeidsform.varierende.info.tittel')}>
-                    <FormattedMessage id="arbeidsforhold.arbeidsform.varierende.info.tekst.1" />
+                <ExpandableInfo title={intlHelper(intl, 'arbeidsforhold.arbeidsform.fast.info.tittel')}>
+                    <FormattedMessage id="arbeidsforhold.arbeidsform.fast.info.tekst" />
                 </ExpandableInfo>
             );
         case Arbeidsform.turnus:
@@ -45,13 +44,28 @@ const ArbeidsformInfoSNFrilanser: React.FunctionComponent<Props> = ({ arbeidsfor
                     </Box>
                 </ExpandableInfo>
             );
-        default:
+        case Arbeidsform.varierende:
             return (
-                <ExpandableInfo title={intlHelper(intl, 'snFrilanser.arbeidsforhold.arbeidsform.fast.info.tittel')}>
-                    <FormattedMessage id="snFrilanser.arbeidsforhold.arbeidsform.fast.info.tekst" />
+                <ExpandableInfo title={intlHelper(intl, 'arbeidsforhold.arbeidsform.varierende.info.tittel')}>
+                    <FormattedMessage id="arbeidsforhold.arbeidsform.varierende.info.tekst.1" />
+                    <Box margin="l">
+                        <Element tag="h3">
+                            <FormattedMessage id="arbeidsforhold.arbeidsform.varierende.info.tekst.eksempel" />
+                        </Element>
+
+                        <p>
+                            <FormattedMessage id="arbeidsforhold.arbeidsform.varierende.info.tekst.2" />
+
+                            <br />
+                            <FormattedMessage id="arbeidsforhold.arbeidsform.varierende.info.tekst.3" />
+                        </p>
+                        <p>
+                            <FormattedMessage id="arbeidsforhold.arbeidsform.varierende.info.tekst.4" />
+                        </p>
+                    </Box>
                 </ExpandableInfo>
             );
     }
 };
 
-export default ArbeidsformInfoSNFrilanser;
+export default ArbeidsformInfo;
